Implemented basic CGB screen functionalities to test with cgb-acid-2 (boot screen doesn't work)

+++ gameboy/ppu/ppuStatesMode1.go
@@ -32,7 +32,7 @@ func (st *vBlank) Init(ppu *PPU) {
 
 		// Frame complete, switch buffers
 		ppu.frontBuffer = ppu.backBuffer
-		ppu.backBuffer = new([FrameHeight][FrameWidth]uint8)
+		ppu.backBuffer = new([FrameHeight][FrameWidth]uint16)
 	}
 }
 func (st *vBlank) Next(ppu *PPU) ppuInternalState {
